Finish FollowTrajectory and Move to RobotContainer
Finish FollowTrajectory based on the FRC Trajectory Tutorial and moved
it to RobotContainer's `getAutoCommand()` method.
